pyproject.toml
[tool.poetry]
name = "matlab-adapter"
version = "0.0.1"
description = ""
authors = ["Katherine Wilsdon <katherine.wilsdon@inl.gov>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
flask = "*"
requests = "*"
python-dotenv = "*"
deep-lynx = "*"
matlabengine = "9.9.1"
scipy = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
